Added support for two-way belongs_to associations.
Now, any belongs_to attribute that isn't actually used for URL
generation is treated as a normal attribute on the object. This allows
you to update multiple foreign keys on an object with a single
request.

This is order-dependent -- associations showing up first in the class
definition will be higher-priority when it comes to selecting which
association to use for url generation.